chore: fix duplicate test file
